# rumicoach-backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: rumi.coach API
  version: 1.0.0
servers:
  - url: /v1
paths:
...
  /me:
    get:
      summary: Get current authenticated user
      tags:
        - Users
      operationId: GetCurrentUser
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
    patch:
      summary: Update current user
      tags:
        - Users
      operationId: UpdateCurrentUser
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        '200':
          description: Updated user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
    delete:
      summary: Delete current user (anonymize account)
      tags:
        - Users
      operationId: DeleteCurrentUser
      security:
        - BearerAuth: []
      responses:
        '204':
          description: Account anonymized successfully
        '401':
          description: Unauthorized
...
  /feedback:
    post:
      summary: Send feedback, a bug report, or a feature request
      description: >
        Stores the report and any attached screenshots, and notifies the team. Images are
        base64-encoded in the body, capped at 3 per report and 5MB each after decoding, and
        their format is sniffed from the bytes rather than taken from the client. An upload
        that fails costs the screenshot, never the report itself.
      tags:
        - Users
      operationId: SubmitFeedback
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeedbackRequest'
      responses:
        '201':
          description: Feedback recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedbackResponse'
        '400':
          description: Invalid category, empty description, or an unacceptable image
        '401':
          description: Unauthorized
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        phoneNumber:
          type: string
        dateOfBirth:
          type: string
          format: date
        gender:
          type: string
        coachGender:
          type: string
        country:
          type: string
        region:
          type: string
        dataRegion:
          type: string
          enum: [eu, us]
        preferredLanguage:
          type: string
        isActive:
          type: boolean
        waitlisted:
          type: boolean
          description: "True for a waitlist signup whose regional profile has not been provisioned yet; activating the user provisions it."
        createdAt:
          type: string
          format: date-time
          description: "Signup time (for waitlisted users, when they joined the waitlist)."
        theme:
          type: string
        focusArea:
          type: string
          nullable: true
          description: "The Wheel of Life area the user committed to work on (chosen during onboarding), in the user's language."
        coachVoice:
          type: string
          enum: [algieba, charon, despina, enceladus, gacrux, orus, vindemiatrix]
        termsAndConditionsAcceptedAt:
          type: string
          format: date-time
        marketingAcceptedAt:
          type: string
          format: date-time
        lastOnlineAt:
          type: string
          format: date-time
          nullable: true
        lastPlatform:
          type: string
          nullable: true
        aiAcceptedAt:
          type: string
          format: date-time
        balanceSeconds:
          type: integer
          format: int64
          description: >-
            Remaining voice-session balance in seconds (may be negative). Starting a
            billable session requires at least 60 seconds — anything below that is
            refused on the session WebSocket, which sends an INSUFFICIENT_BALANCE error
            frame and then closes with code 4402.
        inFirstJourney:
          type: boolean
          description: >-
            True while the user still has free introductory sessions left — the onboarding
            intro and the Vision session it hands over to — and is therefore exempt from the
            balance gate. Derived from what those sessions produce: the profile details the
            intro collects (country, date of birth, gender) and the ideal-life vision the
            Vision session writes. It stays true while either is still missing, so a session
            the user abandoned after a few seconds costs them nothing.

            Informational only. Clients must NOT gate on it: the WebSocket is the
            enforcement and it now refuses in a form a client can read, so an app that
            pre-empts the call is keeping a second copy of a billing rule that will drift
            from this one. The name predates the change of rule and is kept because shipped
            app builds read it.
        chatHistoryRetentionDays:
          type: integer
          description: How long the WhatsApp/Telegram conversation is kept, in days; 0 keeps it indefinitely.
...
    UserUpdateRequest:
      type: object
      properties:
        name:
          type: string
        dateOfBirth:
          type: string
          format: date
        gender:
          type: string
        theme:
          type: string
        coachGender:
          type: string
        country:
          type: string
        region:
          type: string
        preferredLanguage:
          type: string
        coachVoice:
          type: string
          enum: [algieba, charon, despina, enceladus, gacrux, orus, vindemiatrix]
        chatHistoryRetentionDays:
          type: integer
          enum: [0, 3, 7, 30]
          description: >
            How long the WhatsApp/Telegram conversation is kept, in days. 0 keeps it
            indefinitely and is the only value that gives up the guarantee, so the app
            should say so rather than list it plainly. Changing this rewrites the expiry of
            the existing conversation, so lowering it takes effect immediately and
            retroactively. Nothing below 3 days is offered: the coach replays the last 72
            hours, and a shorter retention would have it losing the thread mid-exchange.
        termsAndConditionsAccepted:
          type: boolean
          description: Consent — see aiAccepted for how all three behave. Required to use the app.
        marketingAccepted:
          type: boolean
          description: Consent — see aiAccepted for how all three behave. The one optional consent.
        aiAccepted:
          type: boolean
          description: >
            Consent to the user's voice being processed by Google's AI. Required to use the
            app: sessions are the product and they cannot run without it.

            These three are collected at registration, but not everyone arrives that way.
            "Continue with Apple" (or Google) on the sign-in screen creates an account
            without passing through the signup steps, so such a user holds none of them —
            not even the Terms — and neither do accounts predating a given consent. Those
            users are asked in the app before their first session and the answers land here,
            which is why all three are writable and not just the AI one.

            Write-only and one-way, all three. Only `true` is meaningful: it stamps the
            matching `*AcceptedAt` with the server's clock, and does so once — a second
            `true` leaves the original moment intact, because when consent was first given
            is the auditable fact. `false` is ignored rather than treated as a withdrawal;
            revoking the required ones means deleting the account, not editing a profile
            field. The timestamps are returned on the User schema as
            `termsAndConditionsAcceptedAt`, `marketingAcceptedAt` and `aiAcceptedAt`.
    FeedbackRequest:
      type: object
      required: [category, description]
      properties:
        category:
          type: string
          enum: [bug, feedback, feature]
        description:
          type: string
          description: The user's own words. Truncated past 8000 characters rather than rejected.
        images:
          type: array
          maxItems: 3
          items:
            type: string
            description: Base64-encoded image, with or without a data-URL prefix.
        context:
          $ref: '#/components/schemas/FeedbackContext'
    FeedbackContext:
      type: object
      description: >
        Diagnostics gathered by the client. Every field is optional — send whatever the
        platform can tell you. A bug report without a version and a screen name costs a
        round trip to the user before anyone can act on it, which is usually a report lost.
      properties:
        appVersion:
          type: string
          description: e.g. "1.4.2"
        buildNumber:
          type: string
        osVersion:
          type: string
          description: e.g. "iOS 17.2", "Android 14", or the browser and version on web.
        deviceModel:
          type: string
          description: e.g. "iPhone15,3". Mobile only.
        locale:
          type: string
          description: The app's active language, which is not always the device's.
        timezone:
          type: string
          description: IANA name, so timestamps in the report can be read in the user's own day.
        screen:
          type: string
          description: Where the user was when they opened the form, e.g. "(tabs)/growth".
        screenSize:
          type: string
          description: e.g. "393x852". Layout bugs are unreproducible without it.
        userAgent:
          type: string
          description: Web only.
    FeedbackResponse:
      type: object
      properties:
        id:
          type: string
...